Focus done trigger to be worked on !
