+++ index.html
@@ -5,6 +5,7 @@
         <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js" integrity="sha384-7+zCNj/IqJ95wo16oMtfsKbZ9ccEh31eOz1HGyDuCQ6wgnyJNSYdrPa03rtR1zdB" crossorigin="anonymous"></script>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
         <link rel="stylesheet" href="./static/styling.css">
+        <script src="https://requirejs.org/docs/release/2.3.5/minified/require.js"></script>
     </head>
     <body>
         <div id="page">
@@ -28,5 +29,5 @@
             </div>
         </div>
     </body> 
-    <script type="text/javascript" src="./timer.js"></script>
+    <script type="module" src="./lib/bundle.js"></script>
 </html>
