+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/amalmajeed/pomodorotimer#readme",
   "dependencies": {
+    "browserify": "^17.0.0",
     "lodash": "^4.17.21"
   }
 }
